# src/main/java/ru/astra/time/enums/ShopUnitType.java
package ru.astra.time.enums;

/**
 * Тип элемента (продукта) - категория или товар.
 */
public enum ShopUnitType {

    /**
     * Категория.
     */
    CATEGORY,

    /**
     * Товар.
     */
    OFFER
}